setup_lnx.sh: Fixed issue with internal being deleted on second run.
The script has the option to do an 'in-place update' of QB64 which
involves deleting the internal folder. This fix keeps the script from
deleting internal when the option is turned off (DOWNLOAD=0). Now the
script will recompile qb64 if run a second time.

+++ setup_lnx.sh
@@ -53,7 +53,7 @@ QB64_ICON_NAME="qb64icon32.png"
 
 DISTRO=
 
-if [ -f ./qb64 ]; then
+if [ -f ./qb64 ] && [ "$DOWNLOAD" == "1" ]; then
   echo "Removing old QB64 files in preperation for installing new version..."
   rm ./qb64
   rm -fr ./internal
